+++ packages/jaspr/lib/src/browser/browser_binding.dart
@@ -9,8 +9,26 @@ import 'dom_render_object.dart';
 
 /// Global component binding for the browser
 class BrowserAppBinding extends AppBinding with ComponentsBinding {
+
   @override
   bool get isClient => true;
+    
+  static final String _baseOrigin = () {
+    var base = document.querySelector('head>base') as HTMLBaseElement?;
+    return base?.href ?? window.location.origin;
+  }();
+
+  @override
+  String get currentUrl {
+    if (_baseOrigin.length > window.location.href.length) {
+      return '/';
+    }
+    var pathWithoutOrigin = window.location.href.substring(_baseOrigin.length);
+    if (!pathWithoutOrigin.startsWith('/')) {
+      pathWithoutOrigin = '/$pathWithoutOrigin';
+    }
+    return pathWithoutOrigin;
+  }
 
   late String attachTarget;
   late (Node, Node)? attachBetween;
