+++ packages/jaspr/lib/src/foundation/binding.dart
@@ -9,6 +9,12 @@ abstract class AppBinding with SchedulerBinding {
   /// Whether the current app is run on the client (in the browser)
   bool get isClient;
 
+  /// The currently active url.
+  /// 
+  /// On the server, this is the request url. 
+  /// On the client, this is the currently visited url in the browser.
+  String get currentUrl;
+
   /// The [Element] that is at the root of the hierarchy.
   ///
   /// This is initialized when [runApp] is called.
